Variable Fixes
Fixed the need for multiple edge names for the nested edge deployment.

Modified it so the edges are deployed with FQDN on the nested edges so it's inline with the non-nested plays

Modified the CEIP call to change the revision number to support 1 call only for a fresh install

Modified requirements.txt for the required packages to deploy this solution.